Refactor notification icon resolution logic for improved clarity and efficiency
