+++ src/modules/notifications.rs
@@ -12,19 +12,22 @@ use crate::{
     theme::AshellTheme,
 };
 use chrono::{DateTime, Local};
+use freedesktop_icons::lookup;
 use iced::{
     Alignment, Background, Border, Color, Element, Length, Radius, Subscription, Task, Theme,
     widget::{
         Space, button, column, container, horizontal_rule, image, row, scrollable, svg, text,
     },
 };
+use linicon_theme::get_icon_theme;
 use log::error;
 use std::{
     collections::{HashMap, HashSet, VecDeque},
-    path::Path,
+    path::{Path, PathBuf},
     time::Duration,
 };
 use zbus::Connection;
+use zbus::zvariant::OwnedValue;
 
 // Constants for UI dimensions and styling
 const ICON_SIZE: f32 = 20.0;
@@ -60,21 +63,104 @@ enum NotificationIcon {
 }
 
 fn resolve_notification_icon(notification: &Notification) -> NotificationIcon {
-    match notification.resolved_icon_path.as_ref() {
+    match resolve_notification_icon_path(
+        &notification.app_name,
+        &notification.app_icon,
+        &notification.hints,
+    ) {
         Some(path) => {
-            let is_svg = Path::new(path)
+            let is_svg = Path::new(&path)
                 .extension()
                 .and_then(|ext| ext.to_str())
                 .is_some_and(|ext| ext.eq_ignore_ascii_case("svg"));
 
             if is_svg {
-                NotificationIcon::Vector(svg::Handle::from_path(path.clone()))
+                NotificationIcon::Vector(svg::Handle::from_path(path))
             } else {
-                NotificationIcon::Raster(image::Handle::from_path(path.clone()))
+                NotificationIcon::Raster(image::Handle::from_path(path))
             }
         }
         None => NotificationIcon::Bell,
     }
+}
+
+fn non_empty_owned_value_string(value: Option<&OwnedValue>) -> Option<String> {
+    value
+        .and_then(|v| v.clone().try_into().ok())
+        .map(|s: String| s.trim().to_string())
+        .filter(|s| !s.is_empty())
+}
+
+fn parse_file_url(value: &str) -> Option<PathBuf> {
+    if !value.starts_with("file://") {
+        return None;
+    }
+
+    let decoded = url::Url::parse(value).ok()?.to_file_path().ok()?;
+    decoded.exists().then_some(decoded)
+}
+
+fn find_icon_path(icon_name: &str) -> Option<PathBuf> {
+    let base_lookup = lookup(icon_name).with_cache();
+
+    match get_icon_theme() {
+        Some(theme) => base_lookup.with_theme(&theme).find().or_else(|| {
+            let fallback_lookup = lookup(icon_name).with_cache();
+            fallback_lookup.find()
+        }),
+        None => base_lookup.find(),
+    }
+}
+
+fn resolve_notification_icon_path(
+    app_name: &str,
+    app_icon: &str,
+    hints: &HashMap<String, OwnedValue>,
+) -> Option<String> {
+    let mut candidates = Vec::new();
+
+    if !app_icon.trim().is_empty() {
+        candidates.push(app_icon.trim().to_string());
+    }
+
+    for key in [
+        "image-path",
+        "image_path",
+        "icon-name",
+        "icon_name",
+        "desktop-entry",
+    ] {
+        if let Some(value) = non_empty_owned_value_string(hints.get(key)) {
+            candidates.push(value);
+        }
+    }
+
+    if !app_name.trim().is_empty() {
+        candidates.push(app_name.trim().to_string());
+    }
+
+    for candidate in candidates {
+        if let Some(path) = parse_file_url(&candidate) {
+            return Some(path.to_string_lossy().into_owned());
+        }
+
+        let candidate_path = PathBuf::from(&candidate);
+        if (candidate.contains('/') || candidate.starts_with('.')) && candidate_path.exists() {
+            return Some(candidate_path.to_string_lossy().into_owned());
+        }
+
+        if let Some(path) = find_icon_path(&candidate) {
+            return Some(path.to_string_lossy().into_owned());
+        }
+
+        if let Some(stripped) = candidate.strip_suffix(".desktop")
+            && let Some(path) = find_icon_path(stripped)
+        {
+            return Some(path.to_string_lossy().into_owned());
+        }
+    }
+
+    None
 }
 
 fn notification_icon_with_frame<'a, M: 'a>(icon_kind: &NotificationIcon) -> Element<'a, M> {
